Use fastfpe
The benchmarks show fastfpe can be 16x faster that
the Python FF3-1. This is a remarkable
improvement, reducing the performance penalty of
the obfuscation process.

While there's no guarantee that the rust AES is as
secure as the pycryptodome one, it's generating
identical output and the stable tests pass. I
think we are more likely to encounter a future
issue with FF3-1 itself -- the curse of small
domains -- than with the AES implementation.